Add files via upload
Additions made to Player class to incorporate a temporary Battle Log for display during combat. 
Static function was added to Player class to allow the comparing of actions. Math and outcomes still need to be calculated.
Match Interface in the process of being prototyped. End condition still needs to be tested.